requirements.txt
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
httpx==0.27.2
click==8.1.7
python-dotenv==1.0.1
